Split the stages of wrapping into their own files.
