+++ src/text/wrap.rs
@@ -1,283 +1,16 @@
 use crate::{
     math::{max_f, min_f},
     scene::SceneTarget,
-    style::{Alignment, EffectiveStyle},
-    text::{font::Font, PreparedLine, PreparedSpan, PreparedText, Text},
+    style::Alignment,
+    text::{PreparedLine, PreparedSpan, PreparedText, Text},
     KludgineResult,
 };
 use approx::relative_eq;
 use futures::future::join_all;
-use rusttype::{GlyphId, PositionedGlyph, Scale};
 
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-enum LexerStatus {
-    /// We have wrapped to a new line
-    AtSpanStart,
-    /// We have received at least one glyph for this word
-    InWord,
-    /// We have encountered a punctuation mark after a word.
-    TrailingPunctuation,
-    /// We have encountered a whitespace or punctuation character
-    Whitespace,
-}
-
-#[derive(Debug)]
-enum Token {
-    EndOfLine(rusttype::VMetrics),
-    Characters(PreparedSpan),
-    Punctuation(PreparedSpan),
-    Whitespace(PreparedSpan),
-}
-
-#[derive(Debug)]
-enum SpanGroup {
-    Spans(Vec<PreparedSpan>),
-    Whitespace(Vec<PreparedSpan>),
-    EndOfLine(rusttype::VMetrics),
-}
-
-impl SpanGroup {
-    fn spans(&self) -> Vec<PreparedSpan> {
-        match self {
-            SpanGroup::Spans(spans) => spans.clone(),
-            SpanGroup::Whitespace(spans) => spans.clone(),
-            SpanGroup::EndOfLine(_) => Vec::default(),
-        }
-    }
-}
-
-#[derive(Default)]
-pub struct Lexer {
-    tokens: Vec<Token>,
-}
-
-struct LexerState<'a> {
-    style: &'a EffectiveStyle,
-    font: &'a Font,
-    glyphs: Vec<PositionedGlyph<'static>>,
-    lexer_state: LexerStatus,
-    last_glyph_id: Option<GlyphId>,
-    caret: f32,
-}
-
-impl<'a> LexerState<'a> {
-    fn new(font: &'a Font, style: &'a EffectiveStyle) -> Self {
-        Self {
-            font,
-            style,
-            lexer_state: LexerStatus::AtSpanStart,
-            glyphs: Default::default(),
-            last_glyph_id: None,
-            caret: 0.,
-        }
-    }
-
-    async fn emit_token_if_needed(&mut self) -> Option<Token> {
-        if self.glyphs.is_empty() {
-            None
-        } else {
-            let current_committed_glyphs = std::mem::take(&mut self.glyphs);
-
-            let metrics = self.font.metrics(self.style.font_size).await;
-            let span = PreparedSpan::new(
-                self.font.clone(),
-                self.style.font_size,
-                self.style.color,
-                self.caret,
-                current_committed_glyphs,
-                metrics,
-            );
-            self.caret = 0.0;
-
-            let token = match self.lexer_state {
-                LexerStatus::AtSpanStart => unreachable!(),
-                LexerStatus::InWord => Token::Characters(span),
-                LexerStatus::TrailingPunctuation => Token::Punctuation(span),
-                LexerStatus::Whitespace => Token::Whitespace(span),
-            };
-            Some(token)
-        }
-    }
-}
-
-impl Lexer {
-    // Text (Vec<Span>) -> Vec<Token{ PreparedSpan, TokenKind }>
-    async fn prepare_spans(
-        mut self,
-        text: &Text,
-        scene: &SceneTarget,
-    ) -> KludgineResult<Vec<Token>> {
-        for span in text.spans.iter() {
-            let font = scene
-                .lookup_font(&span.style.font_family, span.style.font_weight)
-                .await?;
-            let vmetrics = font.metrics(span.style.font_size).await;
-
-            let mut state = LexerState::new(&font, &span.style);
-
-            for c in span.text.chars() {
-                if c.is_control() {
-                    if c == '\n' {
-                        self.tokens.push(Token::EndOfLine(vmetrics));
-                    }
-                } else {
-                    let new_lexer_state = if c.is_whitespace() {
-                        LexerStatus::Whitespace
-                    } else if c.is_ascii_punctuation() {
-                        LexerStatus::TrailingPunctuation
-                    } else {
-                        LexerStatus::InWord
-                    };
-
-                    if new_lexer_state != state.lexer_state {
-                        if let Some(token) = state.emit_token_if_needed().await {
-                            self.tokens.push(token);
-                        }
-                    }
-
-                    state.lexer_state = new_lexer_state;
-
-                    let base_glyph = font.glyph(c).await;
-                    if let Some(id) = state.last_glyph_id.take() {
-                        state.caret += font
-                            .pair_kerning(span.style.font_size, id, base_glyph.id())
-                            .await;
-                    }
-                    state.last_glyph_id = Some(base_glyph.id());
-                    let glyph = base_glyph
-                        .scaled(Scale::uniform(span.style.font_size))
-                        .positioned(rusttype::point(state.caret, 0.0));
-
-                    state.caret += glyph.unpositioned().h_metrics().advance_width;
-                    state.glyphs.push(glyph);
-                }
-            }
-
-            if let Some(token) = state.emit_token_if_needed().await {
-                self.tokens.push(token);
-            }
-        }
-
-        Ok(self.tokens)
-    }
-}
-
-pub struct MeasuredText {
-    groups: Vec<SpanGroup>,
-}
-
-struct TextMeasureState {
-    current_group: Option<SpanGroup>,
-    status: ParserStatus,
-    groups: Vec<SpanGroup>,
-}
-
-impl TextMeasureState {
-    fn push_token(&mut self, token: Token) {
-        match token {
-            Token::EndOfLine(vmetrics) => {
-                self.commit_current_group();
-                self.groups.push(SpanGroup::EndOfLine(vmetrics));
-                self.status = ParserStatus::LineStart;
-            }
-            Token::Characters(span) => {
-                match self.status {
-                    ParserStatus::LineStart | ParserStatus::InWord => {
-                        self.push_visual_span(span);
-                        self.status = ParserStatus::InWord;
-                    }
-
-                    ParserStatus::Whitespace | ParserStatus::TrailingPunctuation => {
-                        self.commit_current_group();
-                        self.push_visual_span(span);
-                        self.status = ParserStatus::TrailingPunctuation;
-                    }
-                };
-            }
-            Token::Punctuation(span) => match self.status {
-                ParserStatus::TrailingPunctuation => {
-                    self.push_visual_span(span);
-                }
-                ParserStatus::LineStart | ParserStatus::InWord => {
-                    self.push_visual_span(span);
-                    self.status = ParserStatus::TrailingPunctuation;
-                }
-                ParserStatus::Whitespace => {
-                    self.commit_current_group();
-                    self.push_visual_span(span);
-                    self.status = ParserStatus::TrailingPunctuation;
-                }
-            },
-            Token::Whitespace(span) => match self.status {
-                ParserStatus::Whitespace => {
-                    self.push_whitespace_span(span);
-                }
-                _ => {
-                    self.commit_current_group();
-                    self.push_whitespace_span(span);
-                    self.status = ParserStatus::Whitespace;
-                }
-            },
-        }
-    }
-
-    fn push_visual_span(&mut self, span: PreparedSpan) {
-        if let Some(SpanGroup::Spans(group)) = &mut self.current_group {
-            group.push(span);
-        } else {
-            self.commit_current_group();
-            self.current_group = Some(SpanGroup::Spans(vec![span]));
-        }
-    }
-
-    fn push_whitespace_span(&mut self, span: PreparedSpan) {
-        if let Some(SpanGroup::Whitespace(group)) = &mut self.current_group {
-            group.push(span);
-        } else {
-            self.commit_current_group();
-            self.current_group = Some(SpanGroup::Whitespace(vec![span]));
-        }
-    }
-
-    fn commit_current_group(&mut self) {
-        if let Some(group) = self.current_group.take() {
-            self.groups.push(group);
-        }
-    }
-
-    fn finish(mut self) -> Vec<SpanGroup> {
-        self.commit_current_group();
-        self.groups
-    }
-}
-
-impl MeasuredText {
-    pub async fn new(text: &Text, scene: &SceneTarget) -> KludgineResult<Self> {
-        let mut measured = Self { groups: Vec::new() };
-
-        measured.measure_text(text, scene).await?;
-
-        //println!("{:#?}", measured.groups);
-        Ok(measured)
-    }
-
-    async fn measure_text(&mut self, text: &Text, scene: &SceneTarget) -> KludgineResult<()> {
-        let mut state = TextMeasureState {
-            current_group: None,
-            status: ParserStatus::LineStart,
-            groups: Vec::new(),
-        };
-
-        // Tokens -> "Words" (groups of characters, and where the breaks would happen)
-        for token in Lexer::default().prepare_spans(text, scene).await? {
-            state.push_token(token);
-        }
-
-        self.groups = state.finish();
-
-        Ok(())
-    }
-}
+mod measured;
+mod tokenizer;
+pub(crate) use self::{measured::*, tokenizer::*};
 
 pub struct TextWrapper {
     options: TextWrap,
@@ -285,7 +18,7 @@ pub struct TextWrapper {
     prepared_text: PreparedText,
 }
 
-enum ParserStatus {
+pub(crate) enum ParserStatus {
     LineStart,
     InWord,
     TrailingPunctuation,
